+++ examples/hallo-world-vue/src/main.js
@@ -1,0 +1,7 @@
+import { createApp } from 'vue';
+import App from './App.vue';
+
+// import highlight CSS theme
+import 'prismjs/themes/prism-tomorrow.min.css';
+
+createApp(App).mount('#app');